refactor(ui): move params state to big file of canvas zod stuff
